Create Database class, update wormholes
