+++ data/wormholes.csv
@@ -18,31 +18,31 @@ L,Bremen,Kallis,,,,,,
 L,Bremen,Nyx,,,,,,
 M,Bremen,Tanga,,,,,,
 L,Bremen,Vega,,,,,,
-,Caliban,Nul,,,,,,
-,Caliban,Oberon,,,,,,
-,Caliban,Orion,,,,,,
-,Caliban,Viking,,,,,,
+S,Caliban,Nul,,,,,,
+M,Caliban,Oberon,,,,,,
+M,Caliban,Orion,,,,,,
+M,Caliban,Viking,,,,,,
 S,Cano,Pyro,,,,,,
 M,Cano,Davien,,,,,,
-,Castra,Hadrian,,,,,,
-,Castra,Nyx,,,,,,
-,Castra,Oso,,,,,,
-,Castra,Oya,,,,,,
+L,Castra,Hadrian,,,,,,
+L,Castra,Nyx,,,,,,
+S,Castra,Oso,,,,,,
+L,Castra,Oya,,,,,,
 L,Castra,Pyro,,,,,,
 M,Cathcart,Davien,,,,,,
 M,Cathcart,Hades,,,,,,
 S,Cathcart,Kilian,,,,,,
 L,Cathcart,Nexus,,,,,,
-,Centauri,Elysium,,,,,,
-,Centauri,Nul,,,,,,
+L,Centauri,Elysium,,,,,,
+L,Centauri,Nul,,,,,,
 L,Charon,Genesis,,,,,,
 M,Charon,Helios,,,,,,
 S,Charon,Kins,,,,,,
 L,Charon,Tyrol,,,,,,
 L,Chronos,Kellog,,,,,,
-,Corel,Geddon,,,,,,
-,Corel,Genesis,,,,,,
-,Corel,Nemo,,,,,,
+L,Corel,Geddon,,,,,,
+L,Corel,Genesis,,,,,,
+L,Corel,Nemo,,,,,,
 L,Croshaw,Ferron,,,,,,
 S,Croshaw,Nul,,,,,,
 L,Croshaw,Rhetor,,,,,,
@@ -50,85 +50,85 @@ L,Croshaw,Sol,,,,,,
 L,Davien,Ferron,,,,,,
 L,Davien,Kilian,,,,,,
 L,Davien,Sol,,,,,,
-,Eealus,La'Uo,,,,,,
-,Eealus,Oya,,,,,,
-,Eealus,Rihlah,,,,,,
-,Eealus,Trise,,,,,,
-,Ellis,Kilian,,,,,,
-,Ellis,Magnus,,,,,,
-,Ellis,Min,,,,,,
-,Ellis,Nexus,,,,,,
-,Ellis,Taranis,,,,,,
-,Elysium,Idris,,,,,,
-,Elysium,Leir,,,,,,
-,Elysium,Vanguard,,,,,,
-,Ferron,Idris,,,,,,
-,Fora,Nemo,,,,,,
-,Fora,Rhetor,,,,,,
-,Garron,Idris,,,,,,
-,Garron,Leir,,,,,,
-,Geddon,Gliese,,,,,,
-,Genesis,Taranis,,,,,,
+L,Eealus,La'Uo,,,,,,
+S,Eealus,Oya,,,,,,
+L,Eealus,Rihlah,,,,,,
+M,Eealus,Trise,,,,,,
+L,Ellis,Kilian,,,,,,
+L,Ellis,Magnus,,,,,,
+S,Ellis,Min,,,,,,
+L,Ellis,Nexus,,,,,,
+M,Ellis,Taranis,,,,,,
+L,Elysium,Idris,,,,,,
+S,Elysium,Leir,,,,,,
+L,Elysium,Vanguard,,,,,,
+M,Ferron,Idris,,,,,,
+L,Fora,Nemo,,,,,,
+M,Fora,Rhetor,,,,,,
+L,Garron,Idris,,,,,,
+S,Garron,Leir,,,,,,
+L,Geddon,Gliese,,,,,,
+M,Genesis,Taranis,,,,,,
 M,Goss,Helios,,,,,,
 L,Goss,Osiris,,,,,,
 L,Goss,Tayac,,,,,,
 M,Goss,Terra,,,,,,
 M,Goss,Tyrol,,,,,,
 L,Gurzil,Hadrian,,,,,,
-,Gurzil,Horus,,,,,,
-,Gurzil,Oya,,,,,,
-,Gurzil,Rihlah,,,,,,
-,Hades,Nemo,,,,,,
-,Hades,Nexus,,,,,,
+M,Gurzil,Horus,,,,,,
+S,Gurzil,Oya,,,,,,
+M,Gurzil,Rihlah,,,,,,
+L,Hades,Nemo,,,,,,
+L,Hades,Nexus,,,,,,
 L,Hadrian,Kiel,,,,,,
-,Hadrian,Oya,,,,,,
+L,Hadrian,Oya,,,,,,
 S,Hadrian,Pyro,,,,,,
 L,Hadrian,Terra,,,,,,
-,Helios,Taranis,,,,,,
-,Helios,Tyrol,,,,,,
-,Horus,Kai'pua,,,,,,
-,Horus,Kiel,,,,,,
-,Horus,Rihlah,,,,,,
-,Idris,Rhetor,,,,,,
-,Kabal,Leir,,,,,,
-,Kai'pua,Rihlah,,,,,,
-,Kallis,Oso,,,,,,
-,Kellog,Odin,,,,,,
-,Kellog,Vector,,,,,,
-,Kellog,Virgil,,,,,,
-,Kins,Kyuk'ya,,,,,,
-,Kyuk'ya,Osiris,,,,,,
-,Kyuk'ya,Yā'mon,,,,,,
-,La'Uo,Trise,,,,,,
-,La'Uo,Tohil,,,,,,
-,Leir,Vanguard,,,,,,
-,Leir,Yulin,,,,,,
+M,Helios,Taranis,,,,,,
+S,Helios,Tyrol,,,,,,
+M,Horus,Kai'pua,,,,,,
+L,Horus,Kiel,,,,,,
+S,Horus,Rihlah,,,,,,
+L,Idris,Rhetor,,,,,,
+M,Kabal,Leir,,,,,,
+L,Kai'pua,Rihlah,,,,,,
+L,Kallis,Oso,,,,,,
+L,Kellog,Odin,,,,,,
+M,Kellog,Vector,,,,,,
+M,Kellog,Virgil,,,,,,
+L,Kins,Kyuk'ya,,,,,,
+L,Kyuk'ya,Osiris,,,,,,
+S,Kyuk'ya,Yā'mon,,,,,,
+L,La'Uo,Trise,,,,,,
+L,La'Uo,Tohil,,,,,,
+S,Leir,Vanguard,,,,,,
+M,Leir,Yulin,,,,,,
 L,Magnus,Ellis,,,,,,
 M,Magnus,Terra,,,,,,
 L,Magnus,Stanton,,,,-62284273.861,23467618.051,20198396.608
-,Min,Nexus,,,,,,
-,Nul,Oberon,,,,,,
-,Nul,Vega,,,,,,
+L,Min,Nexus,,,,,,
+L,Nul,Oberon,,,,,,
+M,Nul,Vega,,,,,,
 L,Nyx,Odin,,,,,,
 M,Nyx,Pyro,,,,,,
 L,Nyx,Tohil,,,,,,
 M,Nyx,Virgil,,,,,,
-,Oberon,Tiber,,,,,,
-,Oberon,Vega,,,,,,
-,Oberon,Virgil,,,,,,
-,Odin,Tanga,,,,,,
-,Orion,Tiber,,,,,,
-,Orion,Viking,,,,,,
-,Orion,Vulture,,,,,,
+M,Oberon,Tiber,,,,,,
+L,Oberon,Vega,,,,,,
+L,Oberon,Virgil,,,,,,
+M,Odin,Tanga,,,,,,
+L,Orion,Tiber,,,,,,
+L,Orion,Viking,,,,,,
+L,Orion,Vulture,,,,,,
 S,Oso,Pyro,,,,,,
 M,Oya,Tohil,,,,,,
 M,Pyro,Stanton,34178642.342,40591999.093,0.000,3310491.640,-27979408.221,-2676285.679
 S,Pyro,Terra,,,,,,
 L,Stanton,Terra,51118221.617,-5269981.303,-4339551.619,,,
-,Tal,Th.us'ūng,,,,,,
+L,Tal,Th.us'ūng,,,,,,
 S,Taranis,Terra,,,,,,
 L,Tayac,Terra,,,,,,
-,Th.us'ūng,Yā'mon,,,,,,
-,Tiber,Vector,,,,,,
-,Tiber,Virgil,,,,,,
+M,Th.us'ūng,Yā'mon,,,,,,
+L,Tiber,Vector,,,,,,
+L,Tiber,Virgil,,,,,,
 L,Vega,Virgil,,,,,,
